+++ apps/coordinator/src/reducers/transactionReducer.js
@@ -283,16 +283,19 @@ function finalizeOutputs(state, action) {
   let unsignedTransaction;
   // First try to build the transaction via PSBT, if that fails (e.g. an input doesn't know about its braid),
   // then try to build it using the old TransactionBuilder plumbing.
+  let unsignedPSBT = "";
   try {
     const args = {
       network: state.network,
       inputs: state.inputs.map(convertLegacyInput),
       outputs: state.outputs.map(convertLegacyOutput),
+      includeGlobalXpubs: true,
     };
     const psbt = getUnsignedMultisigPsbtV0(args);
     unsignedTransaction = Transaction.fromHex(
       psbt.data.globalMap.unsignedTx.toBuffer().toString("hex"),
     );
+    unsignedPSBT = psbt.toBase64();
   } catch (e) {
     // probably has an input that isn't braid aware.
     // NOTE: This won't work for txs with taproot outputs
@@ -304,7 +307,7 @@ function finalizeOutputs(state, action) {
   }
   return {
     ...state,
-    ...{ finalizedOutputs: action.value, unsignedTransaction },
+    ...{ finalizedOutputs: action.value, unsignedTransaction, unsignedPSBT },
   };
 }