+++ packages/caravan-clients/src/client.ts
@@ -10,6 +10,7 @@ import {
   bitcoindSendRawTransaction,
   isWalletAddressNotFoundError,
   callBitcoind,
+  bitcoindRawTxData,
 } from "./bitcoind";
 import {
   bitcoindGetAddressStatus,
@@ -18,6 +19,7 @@ import {
   bitcoindWalletInfo,
 } from "./wallet";
 import BigNumber from "bignumber.js";
+import { FeeRatePercentile, Transaction, UTXO } from "./types";
 
 export class BlockchainClientError extends Error {
   constructor(message) {
@@ -26,21 +28,12 @@ export class BlockchainClientError extends Error {
   }
 }
 
-export interface UTXO {
-  txid: string;
-  vout: number;
-  value: number;
-  status: {
-    confirmed: boolean;
-    block_time: number;
-  };
-}
-
 export enum ClientType {
   PRIVATE = "private",
   BLOCKSTREAM = "blockstream",
   MEMPOOL = "mempool",
 }
+
 const delay = () => {
   return new Promise((resolve) => setTimeout(resolve, 500));
 };
@@ -168,6 +161,98 @@ export class BlockchainClient extends ClientBase {
     } catch (error: any) {
       throw new Error(
         `Failed to get UTXOs for address ${address}: ${error.message}`,
+      );
+    }
+  }
+
+  public async getAddressTransactions(address: string): Promise<Transaction[]> {
+    try {
+      if (this.type === ClientType.PRIVATE) {
+        const data = await callBitcoind(
+          this.bitcoindParams.url,
+          this.bitcoindParams.auth,
+          "listtransactions",
+          [this.bitcoindParams.walletName],
+        );
+
+        const txs: Transaction[] = [];
+        for (const tx of data) {
+          if (tx.address === address) {
+            const rawTxData = await bitcoindRawTxData(tx.txid);
+            const transaction: Transaction = {
+              txid: tx.txid,
+              vin: [],
+              vout: [],
+              size: rawTxData.size,
+              weight: rawTxData.weight,
+              fee: tx.fee,
+              isSend: tx.category === "send" ? true : false,
+              amount: tx.amount,
+              block_time: tx.blocktime,
+            };
+            for (const input of rawTxData.vin) {
+              transaction.vin.push({
+                prevTxId: input.txid,
+                vout: input.vout,
+                sequence: input.sequence,
+              });
+            }
+            for (const output of rawTxData.vout) {
+              transaction.vout.push({
+                scriptPubkeyHex: output.scriptPubKey.hex,
+                scriptPubkeyAddress: output.scriptPubKey.address,
+                value: output.value,
+              });
+            }
+            txs.push(transaction);
+          }
+        }
+        return txs;
+      }
+
+      // For Mempool and Blockstream
+      const data = await this.Get(`/address/${address}/txs`);
+      const txs: Transaction[] = [];
+      for (const tx of data.txs) {
+        const transaction: Transaction = {
+          txid: tx.txid,
+          vin: [],
+          vout: [],
+          size: tx.size,
+          weight: tx.weight,
+          fee: tx.fee,
+          isSend: false,
+          amount: 0,
+          block_time: tx.status.block_time,
+        };
+
+        for (const input of tx.vin) {
+          if (input.prevout.scriptpubkey_address === address) {
+            transaction.isSend = true;
+          }
+          transaction.vin.push({
+            prevTxId: input.txid,
+            vout: input.vout,
+            sequence: input.sequence,
+          });
+        }
+
+        let total_amount = 0;
+        for (const output of tx.vout) {
+          total_amount += output.value;
+          transaction.vout.push({
+            scriptPubkeyHex: output.scriptpubkey,
+            scriptPubkeyAddress: output.scriptpubkey_address,
+            value: output.value,
+          });
+        }
+        transaction.amount = total_amount;
+        txs.push(transaction);
+      }
+      return txs;
+    } catch (error: any) {
+      throw new Error(
+        `Failed to get transactions for address ${address}: ${error.message}`,
       );
     }
   }
@@ -312,6 +397,44 @@ export class BlockchainClient extends ClientBase {
       }
     } catch (error: any) {
       throw new Error(`Failed to get fee estimate: ${error.message}`);
+    }
+  }
+
+  public async getBlockFeeRatePercentileHistory(): Promise<
+    FeeRatePercentile[]
+  > {
+    try {
+      if (
+        this.type === ClientType.PRIVATE ||
+        this.type === ClientType.BLOCKSTREAM
+      ) {
+        throw new Error(
+          "Not supported for private clients and blockstream. Currently only supported for mempool",
+        );
+      }
+
+      const data = await this.Get(`/v1/mining/blocks/fee-rates/all`);
+
+      const feeRatePercentileBlocks: FeeRatePercentile[] = [];
+      for (const block of data) {
+        const feeRatePercentile: FeeRatePercentile = {
+          avgHeight: block?.avgHeight,
+          timestamp: block?.timestamp,
+          avgFee_0: block?.avgFee_0,
+          avgFee_10: block?.avgFee_10,
+          avgFee_25: block?.avgFee_25,
+          avgFee_50: block?.avgFee_50,
+          avgFee_75: block?.avgFee_75,
+          avgFee_90: block?.avgFee_90,
+          avgFee_100: block?.avgFee_100,
+        };
+        feeRatePercentileBlocks.push(feeRatePercentile);
+      }
+      return feeRatePercentileBlocks;
+    } catch (error: any) {
+      throw new Error(
+        `Failed to get feerate percentile block: ${error.message}`,
+      );
     }
   }