+++ packages/caravan-health/.eslintrc.js
@@ -1,0 +1,11 @@
+// .eslintrc.js
+module.exports = {
+    root: true,
+    extends: [
+        "@caravan/eslint-config/library.js"
+    ],
+    parser: "@typescript-eslint/parser",
+    parserOptions: {
+        project: true,
+    },
+};